CORE-6364 Cleaning up merkle tree API
