+++ crypto-extensions/src/main/kotlin/net/corda/v5/crypto/extensions/merkle/MerkleTreeHashDigestProvider.kt
@@ -1,7 +1,9 @@
-package net.corda.v5.crypto.merkle
+package net.corda.v5.crypto.extensions.merkle
 
-import net.corda.v5.crypto.DigestAlgorithmName
 import net.corda.v5.crypto.SecureHash
+import net.corda.v5.crypto.merkle.MerkleProof
+import net.corda.v5.crypto.merkle.MerkleTree
+import net.corda.v5.crypto.merkle.MerkleTreeHashDigestFactory
 
 /**
  * Different use cases require different [MerkleTree] calculations.
@@ -14,9 +16,7 @@ import net.corda.v5.crypto.SecureHash
  *  @property digestAlgorithmName Specifies the digest algorithm.
  */
 
-interface MerkleTreeHashDigestProvider {
-    val digestAlgorithmName: DigestAlgorithmName
-
+interface MerkleTreeHashDigestProvider : MerkleTreeHashDigestFactory  {
     /**
      * Calculates the nonce for a leaf.
      * @param index The leaf's index.
