+++ crypto/src/main/kotlin/net/corda/v5/crypto/merkle/HashDigestConstants.kt
@@ -1,10 +1,6 @@
 @file:JvmName("HashDigestConstants")
 package net.corda.v5.crypto.merkle
 
-import net.corda.v5.crypto.DigestAlgorithmName
-import net.corda.v5.base.annotations.DoNotImplement
-import net.corda.v5.base.annotations.Suspendable
-
 /**
  * An implementation of rfc6962 compatible merkle tree
  */
@@ -54,39 +50,3 @@ const val HASH_DIGEST_PROVIDER_NODE_PREFIX_OPTION = "nodePrefix"
  * Nonce hash digest providers require a 32 bytes long ByteArray entropy option.
  */
 const val HASH_DIGEST_PROVIDER_ENTROPY_OPTION = "entropy"
-
-/**
- * [MerkleTreeFactory] creates [MerkleTree]s and [MerkleTreeHashDigestProvider]s.
- */
-@DoNotImplement
-interface MerkleTreeFactory {
-    /**
-     * Creates a [MerkleTree]
-     *
-     * @param leaves The leaves of the tree.
-     * @param digestProvider Merkle Tree Hash digest provider used to construct the tree's node and leaf hashes.
-     *
-     * @return A new [MerkleTree] instance.
-     */
-    @Suspendable
-    fun createTree(
-        leaves: List<ByteArray>,
-        digestProvider: MerkleTreeHashDigestProvider
-    ) : MerkleTree
-
-    /**
-     * Creates a [MerkleTreeHashDigestProvider].
-     *
-     * @param merkleTreeHashDigestProviderName name of the Hash Digest Provider class
-     * @param digestAlgorithmName name of the base Hash algorithm
-     * @param options Hash Digest provider specific options
-     *
-     * @return A new [MerkleTreeHashDigestProvider] instance.
-     */
-    @Suspendable
-    fun createHashDigestProvider(
-        merkleTreeHashDigestProviderName: String,
-        digestAlgorithmName: DigestAlgorithmName,
-        options: Map<String, Any> = emptyMap(),
-    ) : MerkleTreeHashDigestProvider
-}
